Moved code into impls, out of amp-a4a
