+++ extensions/amp-a4a/0.1/amp-a4a.js
@@ -24,8 +24,7 @@ import {
   removeChildren,
   createElementWithAttributes,
 } from '../../../src/dom';
-import {CommonSignals} from '../../../src/common-signals';
-import {analyticsForDoc} from '../../../src/analytics';
+
 import {cancellation, isCancellation} from '../../../src/error';
 import {
   installAnchorClickInterceptor,
@@ -1130,48 +1129,7 @@ export class AmpA4A extends AMP.BaseElement {
         }
       });
     }
-    analyticsForDoc(this.getAmpDoc(), true).then(analytics => {
-      const vis = analytics.getAnalyticsRoot(this.element).getVisibilityManager();
-      // Can be any promise or `null`.
-      const readyPromise = Promise.race([
-        this.signals().whenSignal(CommonSignals.INI_LOAD),
-        this.signals().whenSignal(CommonSignals.LOAD_END),
-      ]);
-      // Element must be an AMP element at this time.
-      debugger;
-      // 50% vis w/o ini load
-      vis.listenElement(this.element, {visiblePercentageMin: 50}, null,
-                        event => {
-                          console.log(event);
-                        });
-      // 50% vis w ini load
-      vis.listenElement(this.element, {visiblePercentageMin: 50, waitFor: 'ini-load'},
-                        this.signals().whenSignal(CommonSignals.INI_LOAD),
-                        event => {
-                          console.log(event);
-                        });
-      // first visible
-      vis.listenElement(this.element, {visiblePercentageMin: 1}, null,
-                        event => {
-                          console.log(event);
-                        });
-      // ini-load
-      vis.listenElement(this.element, {waitFor: 'ini-load'}, this.signals().whenSignal(CommonSignals.INI_LOAD),
-                        event => {
-                          console.log(event);
-                        });
-      // load
-      vis.listenElement(this.element, {waitFor: 'load'}, this.signals().whenSignal(CommonSignals.LOAD_END),
-                        event => {
-                          console.log(event);
-                        });
-      // 50% vis, ini-load and 1 sec
-      vis.listenElement(this.element, {visiblePercentageMin: 1, waitFor: 'ini-load', totalTimeMin: 1},
-                        this.signals().whenSignal(CommonSignals.INI_LOAD),
-                        event => {
-                          console.log(event);
-                        });
-    });
+
     return installFriendlyIframeEmbed(
         this.iframe, this.element, {
           host: this.element,
