+++ extensions/amp-ad-network-adsense-impl/0.1/amp-ad-network-adsense-impl.js
@@ -31,6 +31,8 @@ import {
   isGoogleAdsA4AValidEnvironment,
   extractAmpAnalyticsConfig,
 } from '../../../ads/google/a4a/utils';
+import {CommonSignals} from '../../../src/common-signals';
+import {analyticsForDoc} from '../../../src/analytics';
 import {
   googleLifecycleReporterFactory,
   setGoogleLifecycleVarsFromHeaders,
@@ -264,6 +266,53 @@ export class AmpAdNetworkAdsenseImpl extends AmpA4A {
     if (this.ampAnalyticsConfig_) {
       insertAnalyticsElement(this.element, this.ampAnalyticsConfig_, true);
     }
+
+    analyticsForDoc(this.getAmpDoc(), true).then(analytics => {
+      const vis = analytics.getAnalyticsRoot(this.element).getVisibilityManager();
+      // Can be any promise or `null`.
+      const readyPromise = Promise.race([
+        this.signals().whenSignal(CommonSignals.INI_LOAD),
+        this.signals().whenSignal(CommonSignals.LOAD_END),
+      ]);
+      // Element must be an AMP element at this time.
+      debugger;
+      // 50% vis w/o ini load
+      vis.listenElement(this.element, {visiblePercentageMin: 50}, null,
+                        event => {
+                          console.log(event);
+                        });
+      // 50% vis w ini load
+      vis.listenElement(this.element,
+                        {visiblePercentageMin: 50, waitFor: 'ini-load'},
+                        this.signals().whenSignal(CommonSignals.INI_LOAD),
+                        event => {
+                          console.log(event);
+                        });
+      // first visible
+      vis.listenElement(this.element, {visiblePercentageMin: 1}, null,
+                        event => {
+                          console.log(event);
+                        });
+      // ini-load
+      vis.listenElement(this.element, {waitFor: 'ini-load'},
+                        this.signals().whenSignal(CommonSignals.INI_LOAD),
+                        event => {
+                          console.log(event);
+                        });
+      // load
+      vis.listenElement(this.element, {waitFor: 'load'},
+                        this.signals().whenSignal(CommonSignals.LOAD_END),
+                        event => {
+                          console.log(event);
+                        });
+      // 50% vis, ini-load and 1 sec
+      vis.listenElement(this.element,
+                        {visiblePercentageMin: 1, waitFor: 'ini-load', totalTimeMin: 1},
+                        this.signals().whenSignal(CommonSignals.INI_LOAD),
+                        event => {
+                          console.log(event);
+                        });
+    });
     setStyles(dev().assertElement(this.iframe), {
       width: `${this.size_.width}px`,
       height: `${this.size_.height}px`,
@@ -272,4 +321,3 @@ export class AmpAdNetworkAdsenseImpl extends AmpA4A {
 }
 
 AMP.registerElement('amp-ad-network-adsense-impl', AmpAdNetworkAdsenseImpl);
-
